Adding user system model and authentication.
Adding internationalization for english and spanish.

+++ database/migrations/2025_10_12_000002_create_users_table.php
@@ -1,31 +1,34 @@
 <?php
 
+use Database\Common\DatabaseConnections as DB_CONN;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
 return new class() extends Migration
 {
-    protected $connection = 'operational';
+    protected $connection = DB_CONN::OPERATIONAL;
 
     /**
      * Run the migrations.
      */
     public function up(): void
     {
-        Schema::connection('operational')->create('users', function (Blueprint $table) {
-            $table->uuid('id')->primary();
-            $table->uuid('tenant_id');
-            $table->string('email')->unique();
-            $table->string('full_name');
-            $table->boolean('enabled')->default(true);
-            $table->timestamps();
+        if (! Schema::connection($this->connection)->hasTable('users')) {
+            Schema::connection($this->connection)->create('users', function (Blueprint $table) {
+                $table->uuid('id')->primary();
+                $table->uuid('tenant_id');
+                $table->string('email')->unique();
+                $table->string('full_name');
+                $table->boolean('enabled')->default(true);
+                $table->timestamps();
 
-            $table->foreign('tenant_id')
-                ->references('id')
-                ->on('tenants')
-                ->onDelete('cascade');
-        });
+                $table->foreign('tenant_id')
+                    ->references('id')
+                    ->on('tenants')
+                    ->onDelete('cascade');
+            });
+        }
     }
 
     /**
@@ -33,6 +36,6 @@ return new class() extends Migration
      */
     public function down(): void
     {
-        Schema::connection('operational')->dropIfExists('users');
+        Schema::connection($this->connection)->dropIfExists('users');
     }
 };
